+++ src/main/java/com/scottkillen/mod/scottstweaks/config/Settings.java
@@ -52,7 +52,7 @@ public enum Settings implements ConfigSyncable
 
     private static String getLocalizedComment(String settingName)
     {
-        return StatCollector.translateToLocal("config." + TheMod.INSTANCE.modID() + ':' + settingName);
+        return StatCollector.translateToLocal("config." + TheMod.MOD_ID + ':' + settingName);
     }
 
     @Override